backend_sophia/cnn_api/requirements.txt
Django>=4.2.0
djangorestframework>=3.16.0
django-cors-headers>=4.9.0
tensorflow>=2.20.0
opencv-python>=4.13.0
numpy>=1.24.0
